+++ src/util/getLargestContainer.js
@@ -1,0 +1,5 @@
+export default function getLargestContainer (config) {
+  const containerBps = config.theme.container.maxWidth
+  const lastKey = [...Object.keys(containerBps)].pop()
+  return containerBps[lastKey]
+}